Add files via upload
This is a C program that operates on medical data to develop a simple pharmacy management system. The program is made on Code Blocks; the program allows the user to build a database for medicine stock in a pharmacy. The program allows the user to: add, delete, update, show, sell, and get information about the medicine in stock.